requirements.txt
# requirements.txt
pandas
jsonschema
requests
plotly
reportlab
streamlit
pyyaml